requirements.txt
Django==1.10.6
django-ckeditor==5.5.0
django-filer==1.3.2
django-js-asset==1.1.0
django-mptt==0.8.7
django-polymorphic==1.3.1
easy-thumbnails==2.4.2
Markdown==2.6.11
Pillow==5.0.0
Unidecode==0.4.21
